+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -8,6 +8,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.product.domain.Product;
 
 @Entity
 public class MenuProduct {
