+++ src/main/java/kitchenpos/product/service/ProductDto.java
@@ -1,8 +1,8 @@
-package kitchenpos.dto;
+package kitchenpos.product.service;
 
 import java.math.BigDecimal;
 import java.util.Objects;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductDto {
