+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -4,10 +4,10 @@ import static java.util.stream.Collectors.toList;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
-import kitchenpos.dto.ProductDto;
-import kitchenpos.mapper.ProductMapper;
+import kitchenpos.product.service.ProductService;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.service.ProductDto;
+import kitchenpos.product.service.ProductMapper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
