+++ src/test/java/kitchenpos/MenuGroupIntegrationTest.java
@@ -7,7 +7,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.dto.MenuGroupDto;
-import kitchenpos.dto.ProductDto;
+import kitchenpos.product.service.ProductDto;
 import kitchenpos.fixture.MenuGroupFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
