+++ src/test/java/kitchenpos/fixture/MenuProductFixture.java
@@ -1,8 +1,7 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.dto.MenuProductDto;
 
 public enum MenuProductFixture {
